Rename FlatRiemannianMetric (-> EuclideanMetric) and EuclideanMetric (-> CanonicalEuclideanMetric)
