+++ geomstats/test_cases/geometry/matrices.py
@@ -5,9 +5,7 @@ from geomstats.geometry.matrices import Matrices
 from geomstats.test.test_case import TestCase
 from geomstats.test.vectorization import generate_vectorization_data
 from geomstats.test_cases.geometry.base import MatrixVectorSpaceTestCase
-from geomstats.test_cases.geometry.flat_riemannian_metric import (
-    FlatRiemannianMetricTestCase,
-)
+from geomstats.test_cases.geometry.euclidean import EuclideanMetricTestCase
 
 
 class MatrixOperationsTestCase(TestCase):
@@ -367,5 +365,5 @@ class MatricesTestCase(MatrixVectorSpaceTestCase):
         self.assertAllClose(point, point_, atol=atol)
 
 
-class MatricesMetricTestCase(FlatRiemannianMetricTestCase):
+class MatricesMetricTestCase(EuclideanMetricTestCase):
     pass
