+++ tests/tests_geomstats/test_geometry/data/euclidean.py
@@ -2,8 +2,8 @@ import geomstats.backend as gs
 from geomstats.test.data import TestData
 
 from .base import VectorSpaceTestData
-from .flat_riemannian_metric import FlatRiemannianMetricTestData
 from .mixins import GroupExpMixinsTestData
+from .riemannian_metric import RiemannianMetricTestData
 
 
 class EuclideanTestData(GroupExpMixinsTestData, VectorSpaceTestData):
@@ -14,7 +14,18 @@ class EuclideanTestData(GroupExpMixinsTestData, VectorSpaceTestData):
         return self.generate_tests([dict()])
 
 
-class EuclideanMetricTestData(FlatRiemannianMetricTestData):
+class EuclideanMetricTestData(RiemannianMetricTestData):
+    fail_for_autodiff_exceptions = False
+    fail_for_not_implemented_errors = False
+
+    def inner_product_derivative_matrix_is_zeros_test_data(self):
+        return self.generate_random_data()
+
+    def christoffels_are_zeros_test_data(self):
+        return self.generate_random_data()
+
+
+class CanonicalEuclideanMetricTestData(EuclideanMetricTestData):
     fail_for_autodiff_exceptions = False
     fail_for_not_implemented_errors = False
 
@@ -22,7 +33,7 @@ class EuclideanMetricTestData(FlatRiemannianMetricTestData):
         return self.generate_random_data()
 
 
-class EuclideanMetric2TestData(TestData):
+class CanonicalEuclideanMetric2TestData(TestData):
     def exp_test_data(self):
         tangent_vec = gs.array([[2.0, 1.0], [-2.0, -4.0], [-5.0, 1.0]])
         base_point = gs.array([[2.0, 10.0], [8.0, -1.0], [-3.0, 6.0]])
