+++ tests/tests_geomstats/test_geometry/data/symmetric_matrices.py
@@ -2,7 +2,7 @@ import geomstats.backend as gs
 from geomstats.test.data import TestData
 
 from .base import LevelSetTestData, MatrixVectorSpaceTestData
-from .euclidean import FlatRiemannianMetricTestData
+from .euclidean import EuclideanMetricTestData
 
 
 class SymmetricMatricesTestData(MatrixVectorSpaceTestData):
@@ -93,6 +93,6 @@ class SymmetricHollowMatricesTestData(MatrixVectorSpaceTestData, LevelSetTestDat
     pass
 
 
-class HollowMatricesPermutationInvariantMetricTestData(FlatRiemannianMetricTestData):
+class HollowMatricesPermutationInvariantMetricTestData(EuclideanMetricTestData):
     fail_for_not_implemented_errors = False
     fail_for_autodiff_exceptions = False
