+++ tests/tests_geomstats/test_geometry/test_euclidean.py
@@ -2,15 +2,18 @@ import random
 
 import pytest
 
-from geomstats.geometry.euclidean import Euclidean
+from geomstats.geometry.euclidean import Euclidean, EuclideanMetric
+from geomstats.geometry.spd_matrices import SPDMatrices
 from geomstats.test.parametrizers import DataBasedParametrizer
 from geomstats.test_cases.geometry.euclidean import (
+    CanonicalEuclideanMetricTestCase,
     EuclideanMetricTestCase,
     EuclideanTestCase,
 )
 
 from .data.euclidean import (
-    EuclideanMetric2TestData,
+    CanonicalEuclideanMetric2TestData,
+    CanonicalEuclideanMetricTestData,
     EuclideanMetricTestData,
     EuclideanTestData,
 )
@@ -22,11 +25,24 @@ class TestEuclidean(EuclideanTestCase, metaclass=DataBasedParametrizer):
 
 
 class TestEuclideanMetric(EuclideanMetricTestCase, metaclass=DataBasedParametrizer):
-    space = Euclidean(dim=random.randint(2, 5))
+    _dim = random.randint(2, 5)
+    space = Euclidean(_dim, equip=False).equip_with_metric(
+        EuclideanMetric,
+        metric_matrix=SPDMatrices(_dim, equip=False).random_point(),
+    )
     testing_data = EuclideanMetricTestData()
 
 
+class TestCanonicalEuclideanMetric(
+    CanonicalEuclideanMetricTestCase, metaclass=DataBasedParametrizer
+):
+    space = Euclidean(dim=random.randint(2, 5))
+    testing_data = CanonicalEuclideanMetricTestData()
+
+
 @pytest.mark.smoke
-class TestEuclideanMetric2(EuclideanMetricTestCase, metaclass=DataBasedParametrizer):
+class TestCanonicalEuclideanMetric2(
+    CanonicalEuclideanMetricTestCase, metaclass=DataBasedParametrizer
+):
     space = Euclidean(dim=2)
-    testing_data = EuclideanMetric2TestData()
+    testing_data = CanonicalEuclideanMetric2TestData()
