+++ tests/tests_geomstats/test_geometry/test_minkowski.py
@@ -4,9 +4,9 @@ import pytest
 
 from geomstats.geometry.minkowski import Minkowski
 from geomstats.test.parametrizers import DataBasedParametrizer
-from geomstats.test_cases.geometry.euclidean import EuclideanTestCase
-from geomstats.test_cases.geometry.flat_riemannian_metric import (
-    FlatRiemannianMetricTestCase,
+from geomstats.test_cases.geometry.euclidean import (
+    EuclideanMetricTestCase,
+    EuclideanTestCase,
 )
 
 from .data.minkowski import (
@@ -23,16 +23,12 @@ class TestMinkowski(EuclideanTestCase, metaclass=DataBasedParametrizer):
     testing_data = MinkowskiTestData()
 
 
-class TestMinkowskiMetric(
-    FlatRiemannianMetricTestCase, metaclass=DataBasedParametrizer
-):
+class TestMinkowskiMetric(EuclideanMetricTestCase, metaclass=DataBasedParametrizer):
     space = Minkowski(dim=random.randint(2, 5))
     testing_data = MinkowskiMetricTestData()
 
 
 @pytest.mark.smoke
-class TestMinkowski2Metric(
-    FlatRiemannianMetricTestCase, metaclass=DataBasedParametrizer
-):
+class TestMinkowski2Metric(EuclideanMetricTestCase, metaclass=DataBasedParametrizer):
     space = Minkowski(dim=2)
     testing_data = Minkowski2MetricTestData()
